// front-end-app/quanly-thuysan/src/app/model/nuoinl.ts
export class NuoiNgLieu {
  $class: string;
  idgdnnl: string;
  sanphamnoi: string;
  sanphamngoai: string;
  nguyenlieu: string;
  ao: string;
  trangtrai: string;
  tgnhangiong: string;
  tgxuatao: string;
  phuongtien: string;
  cscb: string;
  soluong: number;
  mota: string;
  trangtrainuoinguyenlieu: string;
}